+++ Source/InputMask/InputMask/Classes/View/MaskedTextInputListener.swift
@@ -3,6 +3,7 @@
 // Created by Jeorge Taflanidi
 //
 
+#if canImport(UIKit) && canImport(Foundation) && !os(watchOS)
 
 import Foundation
 import UIKit
@@ -10,7 +11,12 @@ import UIKit
 
 @available(iOS 11, *)
 public protocol OnMaskedTextChangedListener: AnyObject {
-    func textInput(_ textInput: UITextInput, didExtractValue: String, didFillMandatoryCharacters: Bool)
+    func textInput(
+        _ textInput: UITextInput,
+        didExtractValue value: String,
+        didFillMandatoryCharacters complete: Bool,
+        didComputeTailPlaceholder tailPlaceholder: String
+    )
 }
 
 
@@ -19,7 +25,10 @@ public protocol OnMaskedTextChangedListener: AnyObject {
 open class MaskedTextInputListener: NSObject {
 
     open weak var listener: OnMaskedTextChangedListener?
-    open var onMaskedTextChangedCallback: ((_ textInput: UITextInput, _ value: String, _ complete: Bool) -> ())?
+    open var onMaskedTextChangedCallback: ((_ textInput: UITextInput, _ value: String, _ complete: Bool, _ tailPlaceholder: String) -> ())?
+  
+    open weak var textFieldDelegate: UITextFieldDelegate?
+    open weak var textViewDelegate: UITextViewDelegate?
 
     @IBInspectable open var primaryMaskFormat:   String
     @IBInspectable open var autocomplete:        Bool
@@ -28,19 +37,19 @@ open class MaskedTextInputListener: NSObject {
     @IBInspectable open var rightToLeft:         Bool
     
     /**
-        Allows input suggestions from keyboard
+     Allows input suggestions from keyboard
      */
     @IBInspectable open var allowSuggestions: Bool
     
     /**
-     Shortly after new text is being pasted from the clipboard, ```UITextInput``` receives a new value for its
-     `selectedTextRange` property from the system. This new range is not consistent with the formatted text and
-     calculated caret position most of the time, yet it's being assigned just after ```set caretPosition``` call.
+     Shortly after new text is being pasted from the clipboard, ``UITextInput`` receives a new value for its
+     ``UITextInput/`selectedTextRange`` property from the system. This new range is not consistent with the formatted text and
+     calculated caret position most of the time, yet it's being assigned just after `set caretPosition` call.
      
      To ensure correct caret position is set, it is assigned asynchronously (presumably after a vanishingly
      small delay), if caret movement is set to be non-atomic.
      
-     Default is ```false```.
+     Default is `false`.
      */
     @IBInspectable open var atomicCaretMovement: Bool = false
 
@@ -61,7 +70,7 @@ open class MaskedTextInputListener: NSObject {
         affineFormats: [String] = [],
         affinityCalculationStrategy: AffinityCalculationStrategy = .wholeString,
         customNotations: [Notation] = [],
-        onMaskedTextChangedCallback: ((_ textInput: UITextInput, _ value: String, _ complete: Bool) -> ())? = nil,
+        onMaskedTextChangedCallback: ((_ textInput: UITextInput, _ value: String, _ complete: Bool, _ tailPlaceholder: String) -> ())? = nil,
         allowSuggestions: Bool = true
     ) {
         self.primaryMaskFormat = primaryFormat
@@ -167,8 +176,9 @@ open class MaskedTextInputListener: NSObject {
         )
 
         field.allText = result.formattedText.string
-        field.caretPosition = result.formattedText.string.distanceFromStartIndex(
-            to: result.formattedText.caretPosition
+        setCaretPosition(
+            result.formattedText.string.distanceFromStartIndex(to: result.formattedText.caretPosition),
+            in: field
         )
 
         notifyOnMaskedTextChangedListeners(forTextInput: field, result: result)
@@ -180,12 +190,11 @@ open class MaskedTextInputListener: NSObject {
         _ textInput: UITextInput,
         isChangingCharactersIn range: NSRange,
         replacementString string: String
-    ) -> Mask.Result {
+    ) -> Result {
         let updatedText: String = replaceCharacters(inText: textInput.allText, range: range, withCharacters: string)
         // https://stackoverflow.com/questions/52131894/shouldchangecharactersin-combined-with-suggested-text
         if (allowSuggestions && string == " " && updatedText == " ") {
-//            TODO:
-//            return true
+            return .fallback
         }
         let isDeletion = isThisDeletion(inRange: range, string: string, field: textInput)
         let useAutocomplete = isDeletion ? false : autocomplete
@@ -201,20 +210,12 @@ open class MaskedTextInputListener: NSObject {
         let result: Mask.Result = mask.apply(toText: text)
         
         textInput.allText = result.formattedText.string
+        setCaretPosition(
+            result.formattedText.string.distanceFromStartIndex(to: result.formattedText.caretPosition),
+            in: textInput
+        )
         
-        if self.atomicCaretMovement {
-            textInput.caretPosition = result.formattedText.string.distanceFromStartIndex(
-                to: result.formattedText.caretPosition
-            )
-        } else {
-            DispatchQueue.main.asyncAfter(deadline: DispatchTime.now()) {
-                textInput.caretPosition = result.formattedText.string.distanceFromStartIndex(
-                    to: result.formattedText.caretPosition
-                )
-            }
-        }
-        
-        return result
+        return .notifyListeners(result: result)
     }
     
     open func isThisDeletion(inRange range: NSRange, string: String, field: UITextInput) -> Bool {
@@ -227,13 +228,36 @@ open class MaskedTextInputListener: NSObject {
     }
     
     open func replaceCharacters(inText text: String, range: NSRange, withCharacters newText: String) -> String {
-        if 0 < range.length {
+        /**
+         https://github.com/RedMadRobot/input-mask-ios/issues/84
+         
+         iOS Undo Manager might not consider text adjustments made by the library.
+         Thus, the `range` might be out of bounds compared to the actual text.
+         */
+        var sanitisedRangeLocation = range.location
+        var sanitisedRangeLength = range.length
+        
+        if text.count < sanitisedRangeLocation {
+            sanitisedRangeLocation = text.count
+        }
+        
+        if text.count < sanitisedRangeLocation + sanitisedRangeLength {
+            sanitisedRangeLength = text.count - sanitisedRangeLocation
+        }
+        
+        let sanitisedRange = NSRange(
+            location: sanitisedRangeLocation,
+            length: sanitisedRangeLength
+        )
+        // end
+        
+        if 0 < sanitisedRange.length {
             let result = NSMutableString(string: text)
-            result.replaceCharacters(in: range, with: newText)
+            result.replaceCharacters(in: sanitisedRange, with: newText)
             return result as String
         } else {
             let result = NSMutableString(string: text)
-            result.insert(newText, at: range.location)
+            result.insert(newText, at: sanitisedRange.location)
             return result as String
         }
     }
@@ -271,8 +295,23 @@ open class MaskedTextInputListener: NSObject {
     }
     
     open func notifyOnMaskedTextChangedListeners(forTextInput textInput: UITextInput, result: Mask.Result) {
-        listener?.textInput(textInput, didExtractValue: result.extractedValue, didFillMandatoryCharacters: result.complete)
-        onMaskedTextChangedCallback?(textInput, result.extractedValue, result.complete)
+        listener?.textInput(
+            textInput,
+            didExtractValue: result.extractedValue,
+            didFillMandatoryCharacters: result.complete,
+            didComputeTailPlaceholder: result.tailPlaceholder
+        )
+        onMaskedTextChangedCallback?(textInput, result.extractedValue, result.complete, result.tailPlaceholder)
+    }
+    
+    open func setCaretPosition(_ position: Int, in textInput: UITextInput) {
+        if atomicCaretMovement {
+            textInput.caretPosition = position
+        } else {
+            DispatchQueue.main.asyncAfter(deadline: DispatchTime.now()) {
+                textInput.caretPosition = position
+            }
+        }
     }
 
     private func maskGetOrCreate(withFormat format: String, customNotations: [Notation]) throws -> Mask {
@@ -286,14 +325,19 @@ open class MaskedTextInputListener: NSObject {
         let mask: Mask
         let affinity: Int
     }
+    
+    public enum Result {
+        case notifyListeners(result: Mask.Result)
+        case fallback
+    }
 
     /**
      Workaround to support Interface Builder delegate outlets.
 
-     Allows assigning ```MaskedTextFieldDelegate.listener``` within the Interface Builder.
+     Allows assigning ``MaskedTextInputListener/listener`` within the Interface Builder.
 
-     Consider using ```MaskedTextFieldDelegate.listener``` property from your source code instead of
-     ```MaskedTextFieldDelegate.delegate``` outlet.
+     Consider using ``MaskedTextInputListener/listener`` property from your source code instead of
+     ``MaskedTextInputListener/delegate`` outlet.
      */
     @IBOutlet public var delegate: NSObject? {
         get {
@@ -313,27 +357,73 @@ open class MaskedTextInputListener: NSObject {
 @available(iOS 11, *)
 extension MaskedTextInputListener: UITextFieldDelegate {
 
-    public func textFieldDidBeginEditing(_ textField: UITextField) {
+    open func textFieldDidBeginEditing(_ textField: UITextField) {
+        textFieldDelegate?.textFieldDidBeginEditing?(textField)
         if autocompleteOnFocus && (textField.text ?? "").isEmpty {
             let result: Mask.Result = put(text: "", into: textField, autocomplete: true)
             notifyOnMaskedTextChangedListeners(forTextInput: textField, result: result)
         }
     }
 
-    public func textField(
+    open func textField(
         _ textField: UITextField,
         shouldChangeCharactersIn range: NSRange,
         replacementString string: String
     ) -> Bool {
-        let result: Mask.Result = textInput(textField, isChangingCharactersIn: range, replacementString: string)
-        notifyOnMaskedTextChangedListeners(forTextInput: textField, result: result)
-        return false
+        // NOTE: lib logic depends on controlling the returned value, so no full control forwarding is allowed here
+        if textFieldDelegate?.textField?(textField, shouldChangeCharactersIn: range, replacementString: string) == false {
+            return false
+        }
+        
+        switch textInput(textField, isChangingCharactersIn: range, replacementString: string) {
+            case .fallback:
+                return true
+            case .notifyListeners(let result):
+                notifyOnMaskedTextChangedListeners(forTextInput: textField, result: result)
+                return false
+        }
     }
 
-    public func textFieldShouldClear(_ textField: UITextField) -> Bool {
+    open func textFieldShouldClear(_ textField: UITextField) -> Bool {
+        if textFieldDelegate?.textFieldShouldClear?(textField) == false {
+            return false
+        }
+
         let result: Mask.Result = put(text: "", into: textField, autocomplete: false)
         notifyOnMaskedTextChangedListeners(forTextInput: textField, result: result)
         return true
+    }
+    
+    // MARK: - Call forwarding
+    
+    public func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
+        return textFieldDelegate?.textFieldShouldBeginEditing?(textField) ?? true
+    }
+    
+    public func textFieldShouldEndEditing(_ textField: UITextField) -> Bool {
+        return textFieldDelegate?.textFieldShouldEndEditing?(textField) ?? true
+    }
+    
+    public func textFieldDidEndEditing(_ textField: UITextField) {
+        textFieldDelegate?.textFieldDidEndEditing?(textField)
+    }
+    
+    public func textFieldDidEndEditing(_ textField: UITextField, reason: UITextField.DidEndEditingReason) {
+        textFieldDelegate?.textFieldDidEndEditing?(textField, reason: reason)
+    }
+    
+    @available(iOS 13.0, tvOS 13.0, *)
+    public func textFieldDidChangeSelection(_ textField: UITextField) {
+        textFieldDelegate?.textFieldDidChangeSelection?(textField)
+    }
+    
+    public func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        return textFieldDelegate?.textFieldShouldReturn?(textField) ?? true
+    }
+    
+    @available(iOS 16.0, tvOS 16.0, *)
+    public func textField(_ textField: UITextField, editMenuForCharactersIn range: NSRange, suggestedActions: [UIMenuElement]) -> UIMenu? {
+        return textFieldDelegate?.textField?(textField, editMenuForCharactersIn: range, suggestedActions: suggestedActions)
     }
 
 }
@@ -342,21 +432,78 @@ extension MaskedTextInputListener: UITextFieldDelegate {
 @available(iOS 11, *)
 extension MaskedTextInputListener: UITextViewDelegate {
 
-    public func textViewDidBeginEditing(_ textView: UITextView) {
+    open func textViewDidBeginEditing(_ textView: UITextView) {
+        textViewDelegate?.textViewDidBeginEditing?(textView)
         if autocompleteOnFocus && textView.text.isEmpty {
             let result: Mask.Result = put(text: "", into: textView, autocomplete: true)
             notifyOnMaskedTextChangedListeners(forTextInput: textView, result: result)
         }
     }
 
-    public func textView(
+    open func textView(
         _ textView: UITextView,
         shouldChangeTextIn range: NSRange,
         replacementText text: String
     ) -> Bool {
-        let result: Mask.Result = textInput(textView, isChangingCharactersIn: range, replacementString: text)
-        notifyOnMaskedTextChangedListeners(forTextInput: textView, result: result)
-        return false
+        // NOTE: lib logic depends on controlling the returned value, so no full control forwarding is allowed here
+        if textViewDelegate?.textView?(textView, shouldChangeTextIn: range, replacementText: text) == false {
+            return false
+        }
+        
+        switch textInput(textView, isChangingCharactersIn: range, replacementString: text) {
+            case .fallback:
+                return true
+            case .notifyListeners(let result):
+                notifyOnMaskedTextChangedListeners(forTextInput: textView, result: result)
+                return false
+        }
+    }
+    
+    // MARK: - Call forwarding
+    
+    public func textViewShouldBeginEditing(_ textView: UITextView) -> Bool {
+        return textViewDelegate?.textViewShouldBeginEditing?(textView) ?? true
+    }
+    
+    public func textViewShouldEndEditing(_ textView: UITextView) -> Bool {
+        return textViewDelegate?.textViewShouldEndEditing?(textView) ?? true
+    }
+    
+    public func textViewDidEndEditing(_ textView: UITextView) {
+        textViewDelegate?.textViewDidEndEditing?(textView)
+    }
+    
+    public func textViewDidChange(_ textView: UITextView) {
+        textViewDelegate?.textViewDidChange?(textView)
+    }
+    
+    public func textViewDidChangeSelection(_ textView: UITextView) {
+        textViewDelegate?.textViewDidChangeSelection?(textView)
+    }
+    
+    public func textView(
+        _ textView: UITextView,
+        shouldInteractWith URL: URL,
+        in characterRange: NSRange,
+        interaction: UITextItemInteraction
+    ) -> Bool {
+        return textViewDelegate?.textView?(textView, shouldInteractWith: URL, in: characterRange, interaction: interaction) ?? true
+    }
+    
+    public func textView(
+        _ textView: UITextView,
+        shouldInteractWith textAttachment: NSTextAttachment,
+        in characterRange: NSRange,
+        interaction: UITextItemInteraction
+    ) -> Bool {
+        return textViewDelegate?.textView?(textView, shouldInteractWith: textAttachment, in: characterRange, interaction: interaction) ?? true
+    }
+    
+    @available(iOS 16.0, tvOS 16.0, *)
+    public func textView(_ textView: UITextView, editMenuForTextIn range: NSRange, suggestedActions: [UIMenuElement]) -> UIMenu? {
+        return textViewDelegate?.textView?(textView, editMenuForTextIn: range, suggestedActions: suggestedActions)
     }
 
 }
+
+#endif